Warn the user when a tracee process is cut down by a fatal signal.
